Refactor: Update Navigation and Introduce DialogScene
This commit introduces several changes to the navigation system, including the addition of `DialogScene` and `DialogStrategy` for displaying dialog-like content within the navigation flow. It also refactors how navigation actions and back stacks are handled, particularly for Nav3.

**Key Changes:**

-   **`DialogScene` and `DialogStrategy` (UIViews):**
    -   Added `DialogScene.kt` which defines `DialogScene` to represent a scene where two entries are displayed simultaneously, typically a main content and a dialog overlay.
    -   `DialogStrategy` is a `SceneStrategy` that activates a `DialogScene` if the top two back stack entries are marked for dialog display.
    -   Introduced `DialogScene.dialog()` helper to mark `NavEntry` metadata for dialog presentation.
-   **Navigation Graph Refactoring (UIViews & kmpuiviews):**
    -   **`Nav3Graph.kt` (UIViews):**
        -   Created a new `Nav3Graph.kt` to define the entry provider for Nav3, separating it from the older `NavGraph.kt`.
        -   Uses new helper functions like `dialogEntry` and `twoPaneEntry` for cleaner route definitions.
        -   The main `NavDisplay` in `NavGraphView.kt` now includes `DialogStrategy` in its chain of scene strategies.
    -   **`NavGraph.kt` (UIViews):**
        -   Removed the Nav3 `entryGraph` related code, as it's now in `Nav3Graph.kt`.
    -   **`NavigationActions.kt` (kmpuiviews):**
        -   Added a new composable function `currentDestination(screen: Screen): Boolean` to the `NavigationActions` interface to check if a given screen is the current top-level destination.
    -   **`Navigation2Actions.kt` (kmpuiviews):**
        -   Implemented the new `currentDestination` function by checking the `NavHostController`'s current destination hierarchy.
    -   **`Navigation3Actions.kt` (kmpuiviews):**
        -   Introduced `TopLevelBackStack` to manage the back stack for Nav3, allowing for better handling of top-level navigation and nested stacks.
        -   `Navigation3Actions` now uses `TopLevelBackStack` instead of a simple `SnapshotStateList<NavKey>`.
        -   Implemented the `currentDestination` function by comparing the screen with the `topLevelKey` of the `TopLevelBackStack`.
-   **UI Updates:**
    -   **`HomeNav.kt` (UIViews):**
        -   Updated to use `TopLevelBackStack` for Nav3.
        -   Removed `currentDestination` parameter from `HomeNavigationBar`, `BottomNav`, and `MultipleActions` as this logic is now encapsulated within `NavigationActions`.
        -   Navigation items now use `navActions.currentDestination(screen)` to determine their selected state.
    -   **`MultipleActions.kt` (UIViews):**
        -   Removed the `currentDestination` parameter.
    -   **`MiddleNavAction.kt` (kmpuiviews):**
        -   `ScreenBottomItem` composables now use `LocalNavActions.current.currentDestination(screen)` to determine selected state, removing the `currentDestination` parameter.
-   **Composition Locals (kmpuiviews):**
    -   `LocalNavController` is no longer provided in `KmpLocalCompositionSetup`.
    -   `KmpLocalCompositionSetup` now accepts `TopLevelBackStack<NavKey>` instead of `SnapshotStateList<NavKey>`.
-   **Other Minor Changes:**
    -   Updated `OtakuMaterialTheme.kt` and `MockData.kt` to use `TopLevelBackStack`.
    -   Updated `DesktopUi.kt` to initialize `TopLevelBackStack` for the Nav3 setup.
