+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/navigation/NavGraph.kt
@@ -4,11 +4,8 @@ import androidx.compose.animation.AnimatedContentTransitionScope
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.adaptive.ExperimentalMaterial3AdaptiveApi
-import androidx.compose.material3.adaptive.navigation3.ListDetailSceneStrategy
 import androidx.compose.material3.windowsizeclass.WindowSizeClass
 import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
@@ -16,10 +13,6 @@ import androidx.navigation.compose.dialog
 import androidx.navigation.navDeepLink
 import androidx.navigation.navigation
 import androidx.navigation.toRoute
-import androidx.navigation3.runtime.EntryProviderBuilder
-import androidx.navigation3.runtime.NavKey
-import androidx.navigation3.runtime.entry
-import androidx.navigation3.runtime.entryProvider
 import com.programmersbox.kmpuiviews.presentation.Screen
 import com.programmersbox.kmpuiviews.presentation.about.AboutLibrariesScreen
 import com.programmersbox.kmpuiviews.presentation.navactions.NavigationActions
@@ -41,7 +34,6 @@ import com.programmersbox.kmpuiviews.utils.composables.sharedelements.animatedSc
 import com.programmersbox.uiviews.BuildConfig
 import com.programmersbox.uiviews.GenericInfo
 import com.programmersbox.uiviews.presentation.DebugView
-import com.programmersbox.uiviews.presentation.TwoPaneScene
 import com.programmersbox.uiviews.presentation.all.AllView
 import com.programmersbox.uiviews.presentation.details.DetailsScreen
 import com.programmersbox.uiviews.presentation.favorite.FavoriteUi
@@ -430,227 +422,3 @@ private fun NavGraphBuilder.settings(
         ExtensionList()
     }
 }
-
-@OptIn(ExperimentalComposeUiApi::class, ExperimentalMaterial3Api::class, ExperimentalAnimationApi::class, ExperimentalFoundationApi::class)
-fun entryGraph(
-    customPreferences: ComposeSettingsDsl,
-    notificationLogo: NotificationLogo,
-    windowSize: WindowSizeClass,
-    genericInfo: GenericInfo,
-    navigationActions: NavigationActions,
-) = entryProvider<NavKey> {
-    entry<Screen.RecentScreen> { RecentView() }
-    entry<Screen.DetailsScreen.Details> {
-        DetailsScreen(
-            logo = notificationLogo,
-            windowSize = windowSize,
-            details = koinViewModel { parametersOf(it) }
-        )
-    }
-
-    entry<Screen.ScanQrCodeScreen> { ScanQrCode() }
-
-    entry<Screen.OnboardingScreen> {
-        OnboardingScreen(
-            navController = LocalNavActions.current,
-            customPreferences = customPreferences
-        )
-    }
-
-    entry<Screen.WebViewScreen> {
-        WebViewScreen(
-            url = it.url
-        )
-    }
-
-    entry<Screen.IncognitoScreen> {
-        IncognitoScreen()
-    }
-
-    entry<Screen.AllScreen> {
-        trackScreen(Screen.AllScreen)
-        AllView(
-            isHorizontal = windowSize.widthSizeClass == WindowWidthSizeClass.Expanded
-        )
-    }
-
-    settingsEntryGraph(
-        customPreferences = customPreferences,
-        notificationLogo = notificationLogo,
-        windowSize = windowSize,
-        genericInfo = genericInfo,
-        navigationActions = navigationActions
-    )
-
-    with(genericInfo) { globalNav3Setup() }
-
-}
-
-@OptIn(ExperimentalComposeUiApi::class, ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class, ExperimentalMaterial3AdaptiveApi::class)
-private fun EntryProviderBuilder<NavKey>.settingsEntryGraph(
-    customPreferences: ComposeSettingsDsl,
-    notificationLogo: NotificationLogo,
-    windowSize: WindowSizeClass,
-    genericInfo: GenericInfo,
-    navigationActions: NavigationActions,
-) {
-    entry<Screen.Settings>(
-        metadata = TwoPaneScene.twoPane()
-    ) {
-        SettingScreen(
-            composeSettingsDsl = customPreferences,
-            notificationClick = navigationActions::notifications,
-            favoritesClick = navigationActions::favorites,
-            historyClick = navigationActions::history,
-            globalSearchClick = navigationActions::globalSearch,
-            listClick = navigationActions::customList,
-            debugMenuClick = navigationActions::debug,
-            extensionClick = navigationActions::extensionList,
-            notificationSettingsClick = navigationActions::notificationsSettings,
-            generalClick = navigationActions::general,
-            otherClick = navigationActions::otherSettings,
-            moreInfoClick = navigationActions::moreInfo,
-            moreSettingsClick = navigationActions::moreSettings,
-            geminiClick = { /*navBackStack.add(Screen.GeminiScreen)*/ },
-            sourcesOrderClick = navigationActions::order,
-            appDownloadsClick = navigationActions::downloadInstall,
-            scanQrCode = navigationActions::scanQrCode,
-        )
-    }
-
-    twoPaneEntry<Screen.WorkerInfoScreen> { WorkerInfoScreen() }
-
-    twoPaneEntry<Screen.OrderScreen> {
-        trackScreen(Screen.OrderScreen)
-        SourceOrderScreen()
-    }
-
-    twoPaneEntry<Screen.NotificationsSettings> {
-        trackScreen(Screen.NotificationsSettings)
-        NotificationSettings()
-    }
-
-    twoPaneEntry<Screen.GeneralSettings> {
-        trackScreen(Screen.GeneralSettings)
-        GeneralSettings(customPreferences.generalSettings)
-    }
-
-    twoPaneEntry<Screen.MoreInfoSettings> {
-        trackScreen(Screen.MoreInfoSettings)
-        InfoSettings(
-            usedLibraryClick = navigationActions::about,
-            onPrereleaseClick = navigationActions::prerelease,
-            onViewAccountInfoClick = navigationActions::accountInfo
-        )
-    }
-
-    entry<Screen.PrereleaseScreen> { PrereleaseScreen() }
-
-    twoPaneEntry<Screen.OtherSettings> {
-        trackScreen(Screen.OtherSettings)
-        PlaySettings(customPreferences.playerSettings)
-    }
-
-    twoPaneEntry<Screen.MoreSettings> {
-        trackScreen(Screen.MoreSettings)
-        MoreSettingsScreen()
-    }
-
-    twoPaneEntry<Screen.HistoryScreen> {
-        trackScreen(Screen.HistoryScreen)
-        HistoryUi()
-    }
-
-    entry<Screen.FavoriteScreen> {
-        trackScreen(Screen.FavoriteScreen)
-        FavoriteUi(
-            isHorizontal = windowSize.widthSizeClass == WindowWidthSizeClass.Expanded
-        )
-    }
-
-    twoPaneEntry<Screen.AboutScreen> {
-        trackScreen(Screen.AboutScreen)
-        AboutLibrariesScreen()
-    }
-
-    entry<Screen.GlobalSearchScreen> {
-        trackScreen("global_search")
-        GlobalSearchView(
-            notificationLogo = notificationLogo,
-            isHorizontal = windowSize.widthSizeClass == WindowWidthSizeClass.Expanded,
-            screen = it
-        )
-    }
-
-    entry<Screen.CustomListScreen>(
-        metadata = TwoPaneScene.twoPane()
-    ) {
-        trackScreen(Screen.CustomListScreen)
-        OtakuListView()
-    }
-
-    twoPaneEntry<Screen.CustomListScreen.CustomListItem> {
-        trackScreen(Screen.CustomListScreen)
-        OtakuCustomListScreenStandAlone(it)
-    }
-
-    entry<Screen.CustomListScreen.DeleteFromList> {
-        trackScreen("delete_from_list")
-        DeleteFromListScreen(
-            deleteFromList = it
-        )
-    }
-
-    entry<Screen.ImportListScreen> {
-        trackScreen("import_list")
-        ImportListScreen()
-    }
-
-    entry<Screen.ImportFullListScreen> {
-        trackScreen("import_full_list")
-        ImportFullListScreen()
-    }
-
-    twoPaneEntry<Screen.NotificationScreen> {
-        trackScreen(Screen.NotificationScreen)
-        NotificationScreen()
-    }
-
-    entry<Screen.ExtensionListScreen> {
-        trackScreen(Screen.ExtensionListScreen)
-        ExtensionList()
-    }
-
-    twoPaneEntry<Screen.AccountInfo> {
-        AccountInfoScreen(
-            profileUrl = koinViewModel<AccountViewModel>().accountInfo?.photoUrl?.toString(),
-        )
-    }
-
-    //additionalSettings()
-
-    entry<Screen.DownloadInstallScreen> {
-        DownloadStateScreen()
-    }
-
-    if (BuildConfig.DEBUG) {
-        entry<Screen.DebugScreen> {
-            DebugView()
-        }
-    }
-
-    with(genericInfo) { settingsNav3Setup() }
-}
-
-private inline fun <reified T : Any> EntryProviderBuilder<*>.twoPaneEntry(
-    noinline content: @Composable (T) -> Unit,
-) = entry<T>(
-    metadata = TwoPaneScene.twoPaneDetails()
-) { content(it) }
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-private inline fun <reified T : Any> EntryProviderBuilder<*>.detailEntry(
-    noinline content: @Composable (T) -> Unit,
-) = entry<T>(
-    metadata = ListDetailSceneStrategy.detailPane()
-) { content(it) }
