+++ UIViews/src/main/java/com/programmersbox/uiviews/theme/OtakuMaterialTheme.kt
@@ -6,8 +6,9 @@ import androidx.compose.material3.MotionScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.navigation.NavHostController
-import androidx.navigation3.runtime.NavBackStack
+import androidx.navigation3.runtime.NavKey
 import com.programmersbox.datastore.NewSettingsHandling
+import com.programmersbox.kmpuiviews.presentation.navactions.TopLevelBackStack
 import com.programmersbox.kmpuiviews.theme.generateColorScheme
 import com.programmersbox.kmpuiviews.utils.KmpLocalCompositionSetup
 import com.programmersbox.uiviews.GenericInfo
@@ -20,7 +21,7 @@ import org.koin.core.annotation.KoinExperimentalAPI
 @Composable
 fun OtakuMaterialTheme(
     navController: NavHostController,
-    navBackStack: NavBackStack,
+    navBackStack: TopLevelBackStack<NavKey>,
     genericInfo: GenericInfo,
     settingsHandling: NewSettingsHandling,
     content: @Composable () -> Unit,
