+++ UIViews/src/main/java/com/programmersbox/uiviews/utils/MockData.kt
@@ -41,7 +41,6 @@ import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
-import androidx.navigation3.runtime.rememberNavBackStack
 import com.programmersbox.datastore.NewSettingsHandling
 import com.programmersbox.datastore.SettingsSerializer
 import com.programmersbox.datastore.createProtobuf
@@ -57,6 +56,7 @@ import com.programmersbox.kmpuiviews.presentation.Screen
 import com.programmersbox.kmpuiviews.presentation.components.M3CoverCard
 import com.programmersbox.kmpuiviews.presentation.components.placeholder.M3PlaceHolderCoverCard
 import com.programmersbox.kmpuiviews.presentation.navactions.NavigationActions
+import com.programmersbox.kmpuiviews.presentation.navactions.TopLevelBackStack
 import com.programmersbox.kmpuiviews.utils.ComponentState
 import com.programmersbox.kmpuiviews.utils.KmpLocalCompositionSetup
 import com.programmersbox.kmpuiviews.utils.LocalNavHostPadding
@@ -248,7 +248,7 @@ fun PreviewTheme(
             ) {
                 KmpLocalCompositionSetup(
                     navController,
-                    rememberNavBackStack(Screen.RecentScreen)
+                    remember { TopLevelBackStack(Screen.RecentScreen) }
                 ) {
                     Surface { content() }
                 }
