+++ kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/utils/CompositionLocals.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.platform.UriHandler
@@ -22,6 +21,7 @@ import com.programmersbox.kmpuiviews.customKamelConfig
 import com.programmersbox.kmpuiviews.presentation.navactions.Navigation2Actions
 import com.programmersbox.kmpuiviews.presentation.navactions.Navigation3Actions
 import com.programmersbox.kmpuiviews.presentation.navactions.NavigationActions
+import com.programmersbox.kmpuiviews.presentation.navactions.TopLevelBackStack
 import com.programmersbox.kmpuiviews.repository.CurrentSourceRepository
 import io.kamel.image.config.LocalKamelConfig
 import kotlinx.datetime.LocalDateTime
@@ -29,10 +29,6 @@ import kotlinx.datetime.format.DateTimeFormat
 import org.koin.compose.koinInject
 
 val LocalNavHostPadding = staticCompositionLocalOf<PaddingValues> { error("") }
-val LocalNavController = staticCompositionLocalOf<NavHostController> { error("No NavController Found!") }
-
-//TODO: Change to NavigationActions
-// Maybe also include one for the actual backstack?
 val LocalNavActions = staticCompositionLocalOf<NavigationActions> { error("No NavController Found!") }
 val LocalItemDao = staticCompositionLocalOf<ItemDao> { error("nothing here") }
 val LocalBlurDao = staticCompositionLocalOf<BlurHashDao> { error("nothing here") }
@@ -46,7 +42,7 @@ val LocalSystemDateTimeFormat = staticCompositionLocalOf<DateTimeFormat<LocalDat
 @Composable
 fun KmpLocalCompositionSetup(
     navController: NavHostController,
-    navBackStack: SnapshotStateList<NavKey>,
+    navBackStack: TopLevelBackStack<NavKey>,
     content: @Composable () -> Unit,
 ) {
     val defaultUriHandler = LocalUriHandler.current
